// oddiville-mobile/src/types/typography.ts
export enum typographyEnum {
    H1 = "H1",
    H2 = "H2",
    H3 = "H3",
    H4 = "H4",
    H5 = "H5",
    H6 = "H6",
    B1 = "B1",
    B2 = "B2",
    B3 = "B3",
    B4 = "B4",
    B5 = "B5",
    B6 = "B6",
    C1 = "C1",
    BoldHeading = "BoldHeading",
    BoldHeadingV2 = "BoldHeadingV2",
    SubHeading = "SubHeading",
    SubHeadingV2 = "SubHeadingV2",
    SubHeadingV3 = "SubHeadingV3"
}